move get_layer
